Remove END_PUSH_PROMISE in favor of END_HEADERS
This simplifies the text and makes flag consistent with CONTINUATION
which only bears END_HEADERS.

Technically, both have the same meaning of ending sequence of
header block fragments. Therefore there is no reason to use the
different flag name for the same purpose.

This will also make the implementaion easier because most likely
HEADERS and PUSH_PROMISE handling share the same code and
dancing with END_PUSH_PROMISE and END_HEADERS is not fun.
Although currently they have the same number, it is not
guaranteed to the last. If they are guarenteed to be the same, then
it is simpler to use the same name for same value.