Added Reset State
The close state was too complex as it had significant conditional
behaviour that depended on how the state was entered, plus unspecified
timeout to terminate that conditional behaviour.   This indicated that
close was actually two states: reset and close.

Further, by making RST_STREAM an acknowledged frame, the timeouts for
transitioning from reset to closed can be avoided.  There is no need
for a indeterminant period as we are not closing the connection so it
is easy for the peers to communicate and gracefully complete the reset.