Clear up HPACK Section 4.1 "Calculating Table Size"
It was not quite clear what exactly was referenced with "as defined in Section 5.2". 
In section 5.2, "String length" is defined as the "number of octets used to encode the string". 
This, however, would contradict the (now deleted) second paragraph which explicitly states that the 
length is to be calculated "without any Huffman encoding applied".

Therefore, the reference was deleted and both paragraphs consolidated into one.