Clarifications for trailers
Make it clearer that midders leads to the message being invalid.  I
hope.  It was already pretty clear.

I did have to fix the definition for malformed requests, which was
missing.

And no mention was made of pseudo-header fields.

This uses new terminology too.  That might save effort in a larger
cleanup.

Closes #780.