Speak more about DoS
After carefully reviewing the Netflix advisories, it was clear that the
problem was that implementations didn't read the original text.

The TCP window reduction attack in CVE-2019-9517 was not mentioned, and
is a clever exploit.  The rest were clearly already documented.

I took the opportunity to restructure things a little bit.  The first
few CVEs were about causing the victim to enqueue a bunch of frames and
then exploit the fact that this queueing was badly inefficient.  Those
got a list of their own.

I purposefully left CVE-2019-9516 out.  Though I am sad that people
didn't free() after malloc(), this is pretty much just a systematized
memory leak and I don't feel like calling that out in an RFC.  I've
cited the disclosure page, so I think we're covered there anyway.

Closes #775.