Rewrite :authority description
I think that this is better, but I'd like input on that.

I've tried to simplify this into 3 chunks:

* requests generated natively in HTTP/2
* translating HTTP/1.1 -> HTTP/2
* translating HTTP/2 -> HTTP/1.1

I realized that absolute-form was underserved here.  I could have
excluded it on the basis that it is special, but the rules seemed fairly
clear - at least regarding :authority, so that went in too.

The text now deals more fully with absolute form.  The previous text
really only hinted at the possibility that there might be two values.
This points to -messaging for what to do about those two values (hint:
:authority wins).

Closes #828.