SETTINGS_ENABLE_PUSH from the server
The previous text implied that a value of 1 was a problem,
which was somewhat in tension with the default being 1 and servers
mostly not sending the setting.  So clearly what was being checked by
client was the frame, not the value.  This changes to checking the
setting as it arrives, changes where that text lives.

Also, this changes from "endpoint" to "client" or "server" as
appropriate to match actual practice.

Closes #793.