Make security consequences clearer
This adds an introductory piece to the security considerations.  It's
not uniformly true that the considerations apply only to TLS-using
variants, but it is true enough.

I had to commandeer #894 for this, but I don't want to push to that
branch as this is much bigger than that.

Closes #892.
Closes #894.

This also addresses a comment Julian made about citing TLS and TLS 1.3,
which might be misleading.  I've added a section reference instead.