Move field validation text around
This just moves the reference to the canonical definition of what is
valid up to the top of the section.  That's the canonical text.

This emphasizes the point that HTTP/2 is placing *additional*
requirements on endpoints with respect to validation.

This does not really fix #902 in the sense that it leaves validation of
DQUOTE and friends to the core semantics implementations.

Note that any "MAY" requirement for rejecting a messaging can have the
desired effect on those generating those messages: if an endpoint puts
DQUOTE in a field name, that message will have little hope of being
successfully handled.

Closes #902.