More on connection-specific headers
The original issue here was concerned with Transfer-Encoding, but this
was not a normative statement, it was a note.  With a forward reference,
this is better.

The section on connection-specific headers probably could have been more
specific.  Now that the -semantics draft has a specific list of things
(plus whatever `Connection` lists), we can lean on that more directly.

Closes #938.