Don't justify lowercase requirement
Just state it. No need to justify it so (especially as this invites
questions about the claimed benefits).

This is less than Julian wanted from #963, but I think that it is
important to include a requirement like this very clearly and it isn't
really said anywhere else.  At least not directly.

Closes #963.