Stricter, clearer rules for Dynamic Table Size Update
As they relate to changes in SETTINGS_HEADER_TABLE_SIZE.

This is the strict, by-the-book interpretation.  That is, if the value
changes, you have to send the update, even if it is pointless (that is,
you aren't changing the size of the table).

Note that I'm eliding the point about multiple instructions, which are
necessary if the decoder reduces the maximum below the current size.
The only nod in that direction is "at least one" and a reference to the
section that talks about the need for multiple updates.

This pulls in more of RFC 7541 than I might have liked, but we have to
deal with the integration somewhere and this seems like a reasonable
place to do that.  It might have been better to put this text in a -bis
revision of RFC 7541, but we decided to leave that document as-is.