Only require the update when the size reduces
Otherwise, the table size remains unchanged.

This is a more compatible option than the previous one and - given
circumstances - more likely to see interoperable deployment.

The problem here is that this contradicts (or could be seen to
contradict) text in RFC 7541, Section 4.2.  Rather than play fast and
loose with that, it's probably best to be clear about it and put that in
the abstract.