Pad hexadecimal literals to even digits
We decided to do this for QUIC and it does improve readability and
consistency.

I did this manually, rather than trying to be clever with regexs.  I'd
appreciate a second pair of eyes.

I searched for `0x[0-9a-f](?:[0-9a-f]{2})*[^0-9a-f]` and found nothing;
same for `0x[0-9a-f]*[A-F][0-9a-f]*` (in case we had stray uppercase
strings).