Run the tool I used to clean RFC 9000 on this one
I had to remove the <bcp14> tags from the BCP 14 text, but the tool
restored them.

Other than that, there is no difference in output in the text files that
this generates, so I'm going to say that this was successful.

The source has a ton of whitespace changes though.