Cleanup XML to match AUTH48
Mostly.

I didn't remove the width and height attributes from the SVG.  Removing
these causes the diagram to be stretched to fit, which is not at all
desirable (especially in the default stylesheet).

I restored Kazuho's name in the priority doc references.

I removed entities and used literals for the two &nbsp; instances in the
document (these are now &#xa0;).

I sorted the xi:include elements for references.

The file we got from the RFC editor included quoted values for URI
schemes and pseudo-header field names in the form: `<tt>":scheme"</tt>`.
This was not consistently applied.  I've gone with what we had:
`"<tt>:scheme</tt>"`.